+++ src/main/java/carracing/exception/CarNameFormatException.java
@@ -1,9 +1,6 @@
-package carracing;
+package carracing.exception;
 
 public class CarNameFormatException extends Exception {
-
-    public CarNameFormatException() {
-    };
 
     public CarNameFormatException(String message) {
         super(message);
