+++ src/main/java/carracing/util/Car.java
@@ -1,4 +1,6 @@
-package carracing;
+package carracing.util;
+
+import carracing.exception.CarNameFormatException;
 
 /*
  * Car
