+++ src/test/java/carracing/InputViewTest.java
@@ -1,5 +1,6 @@
 package carracing;
 
+import carracing.view.InputView;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
